+++ translations/en.js
@@ -1,0 +1,12 @@
+"use strict";
+
+$_eseecode.ui.translation = {
+	id: "en",
+	name: "English",
+	code: "en",
+	translator: "Jacobo Vilella",
+	translatorLink: "",
+	strings: {
+	}
+}
+